refactor(types): rename events/agents.ts to agent.ts for naming consistency
Every other file in packages/types/src/events/ is singular (app.ts,
assistant.ts, channel.ts, user.ts, …), keyed to the event-name prefix.
agents.ts was the lone plural; agent.ts matches the agent_* event prefix
and the assistant.ts sibling it was modeled on. Pure rename + index.ts
export update — no type changes.
