+++ packages/types/src/events/index.ts
@@ -1,4 +1,4 @@
-import type { AgentSessionStoppedEvent, AgentSessionTitleChangedEvent } from './agents';
+import type { AgentSessionStoppedEvent, AgentSessionTitleChangedEvent } from './agent';
 import type {
   AppContextChangedEvent,
   AppDeletedEvent,
@@ -89,7 +89,7 @@ import type {
 import type { TokensRevokedEvent } from './token';
 import type { UserChangeEvent, UserHuddleChangedEvent, UserProfileChangedEvent, UserStatusChangedEvent } from './user';
 
-export * from './agents';
+export * from './agent';
 export * from './app';
 export * from './assistant';
 export * from './call';
